add support for content ticks (#113)
One problem we encountered is that an untrusted template could render for a very long time.
For example someone could harm by running `for i in (1..[VERY_LARGE])`

When Liquid starts to render, there is no way to stop it. This introduces something similar like the php tick function for Liquid. This would allow us track the rendering cost at runtime, abort it or do something else while the rendering started.

Co-Authored-By: timglabisch <tim@ainfach.de>